+++ Sources/MapboxMaps/Camera/AnimationOwner.swift
@@ -8,5 +8,8 @@ public struct AnimationOwner: RawRepresentable, Equatable {
     }
 
     public static let gestures = AnimationOwner(rawValue: "com.mapbox.maps.gestures")
+
     public static let unspecified = AnimationOwner(rawValue: "com.mapbox.maps.unspecified")
+
+    internal static let cameraAnimationsManager = AnimationOwner(rawValue: "com.mapbox.maps.cameraAnimationsManager")
 }